Rewrote the delete_domain commcand to use Domain.delete()
